[권한 테이블 수정 및 각 권한에따른 데이터 접근 구현]

+++ src/controllers/calendarControllers/calendar/deleteCalendar.ts
@@ -8,37 +8,31 @@ export default async (req: Request, res: Response) => {
   const { userId, calendarId } = req.body as IUser;
 
   try {
-    const _authority = await getRepository(User)
+    const _user = await getRepository(User)
       .createQueryBuilder('user')
-      .leftJoinAndSelect(
-        'user.userCalendarAuthorities',
-        'userCalendarAuthorities'
-      )
-      .innerJoinAndSelect(
-        'userCalendarAuthorities.calenderAuthority',
-        'calenderAuthority'
-      )
+      .leftJoinAndSelect('user.CalendarAuthorities', 'CalendarAuthorities')
+      .leftJoinAndSelect('CalendarAuthorities.calendar', 'calendar')
       .where('user.id= :userId', { userId })
-      .andWhere('calenderAuthority.calendarId = :calendarId', {
+      .andWhere('CalendarAuthorities.calendarId = :calendarId', {
         calendarId,
       })
       .getOne();
 
-    if (!_authority) {
-      return res.status(400).send('유저 정보 없음 또는 권한이 없는 캘린더');
+    if (_user) {
+      if (_user.CalendarAuthorities[0].write) {
+        await getRepository(Calendar)
+          .createQueryBuilder('calendar')
+          .delete()
+          .where('calendar.id= :calendarId', { calendarId })
+          .execute();
+
+        return res.status(200).send('캘린더 삭제 완료');
+      } else {
+        return res.status(400).send('쓰기 권한 없음');
+      }
+    } else {
+      return res.status(400).send('유저 정보 없음 또는 권한 없는 캘린더');
     }
-  } catch (error) {
-    return res.status(400).send(error);
-  }
-
-  try {
-    await getRepository(Calendar)
-      .createQueryBuilder('calendar')
-      .delete()
-      .where('calendar.id= :calendarId', { calendarId })
-      .execute();
-
-    return res.status(200).send('캘린더 삭제 완료');
   } catch (error) {
     return res.status(400).send(error);
   }
